+++ HaishinKit/Sources/Mixer/MediaMixerOutput.swift
@@ -10,6 +10,13 @@ public protocol MediaMixerOutput: AnyObject, Sendable {
     func mixer(_ mixer: MediaMixer, didOutput sampleBuffer: CMSampleBuffer)
     /// Tells the receiver to an audio buffer incoming.
     func mixer(_ mixer: MediaMixer, didOutput buffer: AVAudioPCMBuffer, when: AVAudioTime)
+    /// Tells the receiver to an audio session event.
+    func mixer(_ mixer: MediaMixer, didReceiveAudioSessionEvent message: String) async
     /// Selects track id for streaming.
     func selectTrack(_ id: UInt8?, mediaType: CMFormatDescription.MediaType) async
+}
+
+public extension MediaMixerOutput {
+    func mixer(_ mixer: MediaMixer, didReceiveAudioSessionEvent message: String) async {
+    }
 }